Add RDT fine-tuning on LIBERO dataset
- Add libero_finetune_correct.py: Correct fine-tuning script based on README guidance
- Add libero_evaluate.py: Model evaluation script for LIBERO tasks
- Add data/hdf5_libero_dataset.py: LIBERO dataset HDF5 loader
- Add eval_sim/eval_rdt_libero.py: LIBERO evaluation simulator
- Add README_LIBERO.md: Detailed usage documentation
- Add LIBERO_FINETUNE_SUMMARY.md: Project summary
- Update configs: Add libero_90 dataset configuration
- Update core files: Fix training and dataset handling
- Clean up: Remove unnecessary files and scripts

+++ models/rdt_runner.py
@@ -42,18 +42,18 @@ class RDTRunner(
             config['lang_adaptor'], 
             in_features=lang_token_dim, 
             out_features=hidden_size
-        )
+        ).to(dtype=dtype)
         self.img_adaptor = self.build_condition_adapter(
             config['img_adaptor'], 
             in_features=img_token_dim, 
             out_features=hidden_size
-        )
+        ).to(dtype=dtype)
         # A `state` refers to an action or a proprioception vector
         self.state_adaptor = self.build_condition_adapter(
             config['state_adaptor'], 
             in_features=state_token_dim * 2,    # state + state mask (indicator)
             out_features=hidden_size
-        )
+        ).to(dtype=dtype)
         
         # Create the noise scheduler
         noise_scheduler_config = config['noise_scheduler']
